# images-api/src/main/java/org/gso/images/controller/ImageController.java
package org.gso.images.controller;

import lombok.RequiredArgsConstructor;
import org.gso.images.dto.ImageResponseDto;
import org.gso.images.service.ImageService;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/v1/images")
@RequiredArgsConstructor
public class ImageController {

    private final ImageService imageService;

    /**
     * GET /api/v1/images/random
     * Public endpoint - returns a random image with optional resizing
     * 
     * @param width - optional width (100-2000)
     * @param height - optional height (100-2000)
     * @param redirect - if true, redirect to resized image
     */
    @GetMapping("/random")
    public ResponseEntity<ImageResponseDto> getRandomImage(
            @RequestParam(required = false) Integer width,
            @RequestParam(required = false) Integer height,
            @RequestParam(defaultValue = "false") Boolean redirect) {
        
        try {
            ImageResponseDto image = imageService.getRandomImage(width, height, redirect);
            
            // If redirect is true and resizedUrl exists, redirect to it
            if (redirect && image.resizedUrl() != null) {
                return ResponseEntity.status(HttpStatus.MOVED_PERMANENTLY)
                        .header("Location", image.resizedUrl())
                        .build();
            }
            
            return ResponseEntity.ok(image);
        } catch (IllegalArgumentException e) {
            return ResponseEntity.badRequest().build();
        }
    }

    /**
     * GET /api/v1/images/resize
     * Resize a given image URL
     * 
     * @param url - the image URL to resize
     * @param width - desired width
     * @param height - desired height
     */
    @GetMapping("/resize")
    public ResponseEntity<ImageResponseDto> resizeImage(
            @RequestParam String url,
            @RequestParam(required = false) Integer width,
            @RequestParam(required = false) Integer height) {
        
        try {
            ImageResponseDto image = imageService.getImageByUrl(url, width, height);
            return ResponseEntity.ok(image);
        } catch (IllegalArgumentException e) {
            return ResponseEntity.badRequest().build();
        }
    }

    /**
     * GET /api/v1/images/list
     * List all available images
     */
    @GetMapping("/list")
    public ResponseEntity<?> listImages() {
        return ResponseEntity.ok(imageService.getAllImages());
    }

    /**
     * Health check endpoint
     */
    @GetMapping("/health")
    public ResponseEntity<String> health() {
        return ResponseEntity.ok("OK");
    }
}